+++ CHANGES.rst
@@ -5,7 +5,9 @@
 1.3.1 (unreleased)
 ==================
 
-- Nothing changed yet.
+- Allow ``IAllowedAttributeProvider`` to be registered to provide
+  additional attributes that would be allowed in sanitized content
+  fragments.
 
 
 1.3.0 (2020-04-06)